+++ packages/components/.storybook/config.js
@@ -4,13 +4,21 @@ import React from 'react';
 
 import { configure, addDecorator } from '@storybook/react';
 import { withA11y } from '@storybook/addon-a11y';
+import { withI18next } from 'storybook-addon-i18next';
+import { locales as tuiLocales } from '@talend/locales-tui/locales';
 
 import 'focus-outline-manager';
-import '../stories/config/i18n';
+import '../../../.storybook/sortStories';
+import i18n from './../../../.storybook/i18n';
 
-function loadStories() {
-	require('../stories');
-}
-
+const languages = {};
+Object.keys(tuiLocales).forEach(key => (languages[key] = key));
+addDecorator(
+	withI18next({
+		i18n,
+		languages,
+	}),
+);
 addDecorator(withA11y);
-configure(loadStories, module);
+
+configure([require.context('../src', true, /\.stories\.js$/)], module);