+++ packages/components/src/TreeView/TreeViewItem/TreeViewItem.component.js
@@ -49,7 +49,7 @@ function TreeViewIcon({ icon, isOpened }) {
 	);
 }
 TreeViewIcon.propTypes = {
-	icon: PropTypes.oneOfType([PropTypes.string, PropTypes.shape(Icon.propTypes)]),
+	icon: PropTypes.oneOfType([PropTypes.bool, PropTypes.string, PropTypes.shape(Icon.propTypes)]),
 	isOpened: PropTypes.bool,
 };
 
@@ -75,7 +75,7 @@ class TreeViewItem extends React.Component {
 			name: PropTypes.string.isRequired,
 			isOpened: PropTypes.bool,
 			children: PropTypes.arrayOf(PropTypes.object),
-			icon: PropTypes.oneOfType([PropTypes.string, PropTypes.object]),
+			icon: PropTypes.oneOfType([PropTypes.bool, PropTypes.string, PropTypes.object]),
 			actions: PropTypes.arrayOf(
 				PropTypes.shape({
 					action: PropTypes.func,
@@ -268,7 +268,7 @@ class TreeViewItem extends React.Component {
 							link
 						/>
 					) : null}
-					<TreeViewIcon key="icon" icon={icon} isOpened={showOpenedFolder} />
+					{icon !== false && <TreeViewIcon key="icon" icon={icon} isOpened={showOpenedFolder} />}
 					<span
 						key="label"
 						className={classNames('tc-treeview-item-name', css['tc-treeview-item-name'])}