+++ packages/logging/src/angular/index.js
@@ -1,3 +1,0 @@
-import initReactLogger from './angular-logger';
-
-export default initReactLogger;